# Conflicts:
#	APP_iOS/APP_iOS/Class/GoodsTableViewCell.m

+++ APP_iOS/APP_iOS/Class/GoodsTableViewCell.m
@@ -26,7 +26,7 @@
     _nameLabel.text = goodModel.name;
     _desLabel.text = goodModel.name;
     _priceLabel.text = [NSString stringWithFormat:@"￥%.1f", goodModel.price];
-    _saleLabel.text = [NSString stringWithFormat:@"销量 %d    库存 %d", goodModel.sales, goodModel.stock];
+    _saleLabel.text = [NSString stringWithFormat:@"销量 %ld    库存 %ld", (long)goodModel.sales, (long)goodModel.stock];
     _timeLabel.text = [NSString stringWithFormat:@"添加时间 %@", [ShareFunction stringWithTimestamp:goodModel.addtime]];
 
 }
